Use Decimal for monetary fields and optimize outstanding report

+++ backend/app/models/order_item.py
@@ -1,5 +1,8 @@
+from decimal import Decimal
+
 from sqlalchemy import BigInteger, DateTime, String, Numeric, ForeignKey, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
+
 from .base import Base
 
 class OrderItem(Base):
@@ -11,5 +14,5 @@ class OrderItem(Base):
     category: Mapped[str | None] = mapped_column(String(50))  # BED | WHEELCHAIR | OXYGEN | ACCESSORY
     item_type: Mapped[str] = mapped_column(String(20))  # OUTRIGHT|INSTALLMENT|RENTAL|FEE
     qty: Mapped[int] = mapped_column(Numeric(12,0), default=1)
-    unit_price: Mapped[float] = mapped_column(Numeric(12,2), default=0)
-    line_total: Mapped[float] = mapped_column(Numeric(12,2), default=0)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
+    line_total: Mapped[Decimal] = mapped_column(Numeric(12, 2), default=0)
